requirements.txt
Django==4.2.3
Pillow==8.3.2
django-compressor
psycopg2-binary
django-nested-admin==3.4.0
django-ckeditor
